+++ config/routes.rb
@@ -122,7 +122,7 @@ Rails.application.routes.draw do
           collection do
             scope module: :tasks do
               resource :analytics, only: :show
-              resource :bulk, only: [:update, :destroy], controller: :bulk
+              resource :bulk, only: [:create, :update, :destroy], controller: :bulk
               resources :filters, only: :index
               resources :tags, only: :index
               namespace :tags do