+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 6 March 2017
+- Added bulk create tasks endpoint
+
 ## 1 March 2017
 - Added ability to sort by start_at field in donations endpoint
 - Added a delete donation endpoint